requirements.txt
tornado==6.0.4
PyJWT==1.7.1
sqlalchemy==1.4.26
PyMySQL==0.9.3
redis==3.5.3
aliyun-python-sdk-ecs==4.19.8
jpush==3.3.8
apscheduler==3.6.3
python-dateutil==2.8.1
aliyun-fc2==2.4.0
pandas==1.0.5
oss2==2.12.1
openpyxl==3.0.4
python-dateutil==2.8.1

pydatahub==2.16.0
urllib3==1.25.9
paho-mqtt==1.5.0
ortools==8.0.8283

pycryptodome==3.10.1
kafka-python==2.0.2
xmltodict==0.12.0
crypto==1.4.1

requests_pkcs12==1.10
pyopenssl==20.0.1
tzlocal==2.1

webargs==8.0.1
swagger-ui-py==21.10.15
apispec==5.1.1
apispec-webframeworks==0.5.2